# docs/swagger.yml
basePath: /api/v1.0
consumes:
- application/json
- application/x-protobuf
definitions:
...
  Server:
    properties:
      address:
        description: Server address
        example:
        - 10.0.0.1/24
        items:
          type: string
        type: array
        x-go-name: Address
      allowedips:
        description: IP addresses allowed to connect
        example:
        - 0.0.0.0/0
        - ::/0
        items:
          type: string
        type: array
        x-go-name: AllowedIPs
      created:
        description: Time when server is created
        example: 26103870
        format: int64
        type: integer
        x-go-name: Created
      dns:
        description: DNS of the VPN server
        example:
        - 1.1.1.1
        items:
          type: string
        type: array
        x-go-name: DNS
      endpoint:
        description: Endpoint of the server
        example: region.example.com
        type: string
        x-go-name: Endpoint
      listenPort:
        description: Port the server listens
        example: 51280
        format: int64
        type: integer
        x-go-name: ListenPort
      mtu:
        format: int64
        type: integer
        x-go-name: Mtu
      persistentKeepalive:
        description: Persistent keep alive for server
        example: 16
        format: int64
        type: integer
        x-go-name: PersistentKeepalive
      postDown:
        description: Post down command
        example: iptables -D FORWARD -i %i -j ACCEPT; iptables -D FORWARD -o %i -j
          ACCEPT; iptables -t nat -D POSTROUTING -o eth0 -j MASQUERADE
        type: string
        x-go-name: PostDown
      postUp:
        description: Post up command
        example: iptables -A FORWARD -i %i -j ACCEPT; iptables -A FORWARD -o %i -j
          ACCEPT; iptables -t nat -A POSTROUTING -o eth0 -j MASQUERADE
        type: string
        x-go-name: PostUp
      preDown:
        description: Pre down command
        example: echo WireGuard PreDown
        type: string
        x-go-name: PreDown
      preUp:
        description: Pre up command
        example: echo WireGuard PreUp
        type: string
        x-go-name: PreUp
      privateKey:
        description: Private key for the server
        example: UFWsgb/Ax5B8zZGx0YtHBAuQVRrOHrxKz2zS2p1LuUE=
        type: string
        x-go-name: PrivateKey
      publicKey:
        description: Public key for the server
        example: T5ZMOnik3YuaRhZgAhcxXrmn2+C0B7qFaqnCypMMcks=
        type: string
        x-go-name: PublicKey
      updated:
        description: Time when server is created
        example: 26103870
        format: int64
        type: integer
        x-go-name: Updated
      updatedBy:
        description: Updater email address
        example: admin@mail.com
        type: string
        x-go-name: UpdatedBy
    type: object
    x-go-package: _/home/sambath/Golang/Revotic-Engineering/erebrus/api/v1/server
...
host: localhost
info:
  contact:
    email: sachinmugu@gmail.com
    name: Sambath Kumar
  description: |-
    Erebrus is an open source VPN solution from The NetSepio, that helps to deploy your own VPN solution in
    minutes.The vision of Erebrus is to deliver Cyber security to everyone .

    Features of Erebrus were, Easy Client and Server management, Supports REST and gRPC, Email VPN configuration to clients easily.

    This documentation guides you, How to use Erebrus endpoints and It's Request and Response briefly.
  license:
    name: GPL-3.0
    url: https://opensource.org/licenses/GPL-3.0
  title: Erebrus
  version: 1.0.0
paths:
...
  /server:
    get:
      description: Retrieves the server details.
      operationId: readServer
      responses:
        "200":
          $ref: '#/responses/serverSucessResponse'
        "400":
          $ref: '#/responses/badRequestResponse'
        "401":
          $ref: '#/responses/unauthorizedResponse'
        "500":
          $ref: '#/responses/serverErrorResponse'
      summary: Read Server
      tags:
      - Server
    patch:
      description: Update the server with given details.
      operationId: updateServer
      parameters:
      - description: Requestbody  used for update server operations.
        in: body
        name: server
        schema:
          $ref: '#/definitions/Server'
        x-go-name: Body
      responses:
        "200":
          $ref: '#/responses/serverSucessResponse'
        "400":
          $ref: '#/responses/badRequestResponse'
        "401":
          $ref: '#/responses/unauthorizedResponse'
        "500":
          $ref: '#/responses/serverErrorResponse'
      summary: Update Server
      tags:
      - Server
  /server/config:
    get:
      description: |-
        Get Server Configuration
        Retrieves the server configuration details.
      operationId: configServer
      responses:
        "200":
          $ref: '#/responses/configResponse'
        "400":
          $ref: '#/responses/badRequestResponse'
        "401":
          $ref: '#/responses/unauthorizedResponse'
        "500":
          $ref: '#/responses/serverErrorResponse'
      tags:
      - Server
...
produces:
- application/json
- application/x-protobuf
- application/config
responses:
  badRequestResponse:
    description: ""
    schema:
      properties:
        Error:
          example: error message
          type: string
        Status:
          example: 400
          format: int64
          type: integer
        Sucess:
          example: false
          type: boolean
      type: object
...
  configResponse:
    description: ""
    schema:
      properties:
        content:
          example: File Download
          type: string
          x-go-name: Data
      type: object
  serverErrorResponse:
    description: ""
    schema:
      properties:
        Error:
          example: error message
          type: string
        Status:
          example: 500
          format: int64
          type: integer
        Sucess:
          example: false
          type: boolean
      type: object
...
  serverSucessResponse:
    description: ""
    schema:
      properties:
        Message:
          example: sucess message
          type: string
        Status:
          example: 201
          format: int64
          type: integer
        Sucess:
          example: true
          type: boolean
        server:
          $ref: '#/definitions/Server'
      type: object
...
  unauthorizedResponse:
    description: ""
    schema:
      properties:
        Error:
          example: error message
          type: string
        Status:
          example: 401
          format: int64
          type: integer
        Sucess:
          example: false
          type: boolean
      type: object
schemes:
- http
- https
swagger: "2.0"